requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
dnspython==2.4.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4
python-dotenv==1.0.0
